+++ admin/live.txt
@@ -9,4 +9,4 @@
 // }
 // ?>
 
- http://localhost:8888/macadams_katrina_portfolio/index.php -->
+ http://localhost:8888/macadams_katrina_portfolio/index.php
