+++ sass/components/_hero.scss
@@ -1,17 +1,12 @@
-
 #hero {
     text-align: left;
     flex-wrap: wrap;
-    justify-content: center;
 }
 
-.hero-text {
-    margin: 0 28px 150px 28px;
-}
 
 #hero h1, h2 {
     margin: 28px 0px;
-    text-shadow: 0 0 9px #ff69a0c3 ;
+    text-shadow: 0 0 9px #ff69a0c3;
 }
 
 #hero h3 {
@@ -23,24 +18,16 @@
     margin: 16px 0;
 }
 
-
-.hero-bg {
-    z-index: -1;
-    top: 30%;
-    position: absolute;
-}
-
 @media screen and (min-width: 744px) {
-
     #hero {
        margin: 112px 28px 24px 28px;
-        
     }
 
-.hero-bg {
-    z-index: -1;
-    top: 20%;
-    position: absolute;
-    height: 500px;
-}
+    .hero-bg {
+        z-index: -1;
+        top: 20%;
+        position: absolute;
+        height: 500px;
+        opacity: 0;
+    }
 }
