+++ sendmail.php
@@ -1,8 +1,4 @@
 <?php
-
-error_reporting(E_ALL);
-ini_set('display_errors', 1);
-
 require_once('includes/connect.php');
 
 header("Access-Control-Allow-Origin: *");
